vec: use uint32_t for array lengths
size_t should be used to represent sizes in bytes and is 64 bits. We
will never have vectors of that amount of elements. Use uint32_t for
array length and capacity.

+++ main/gr_vec.h
@@ -9,8 +9,8 @@
 
 // (internal) vector header
 struct __gr_vec_hdr {
-	size_t len;
-	size_t cap;
+	uint32_t len;
+	uint32_t cap;
 };
 
 // (internal) get a pointer the vector header
@@ -21,22 +21,22 @@ static inline struct __gr_vec_hdr *__gr_vec_hdr(const void *vec) {
 }
 
 // (internal) get the current capacity of a vector
-static inline size_t __gr_vec_cap(const void *vec) {
+static inline uint32_t __gr_vec_cap(const void *vec) {
 	if (vec == NULL)
 		return 0;
 	return __gr_vec_hdr(vec)->cap;
 }
 
 // get the size of a vector
-static inline size_t gr_vec_len(const void *vec) {
+static inline uint32_t gr_vec_len(const void *vec) {
 	if (vec == NULL)
 		return 0;
 	return __gr_vec_hdr(vec)->len;
 }
 
 // (internal) allocate memory to prepare for more items
-static inline void *__gr_vec_grow(void *vec, size_t item_size, size_t add_num, size_t min_cap) {
-	size_t min_len = gr_vec_len(vec) + add_num;
+static inline void *__gr_vec_grow(void *vec, size_t item_size, uint32_t add_num, uint32_t min_cap) {
+	uint32_t min_len = gr_vec_len(vec) + add_num;
 	struct __gr_vec_hdr *hdr;
 
 	if (min_len > min_cap)
@@ -62,9 +62,9 @@ static inline void *__gr_vec_grow(void *vec, size_t item_size, size_t add_num, s
 }
 
 // (internal) delete multiple items stating at a given index
-static inline void __gr_vec_del_range(void *vec, size_t item_size, size_t start, size_t len) {
+static inline void __gr_vec_del_range(void *vec, size_t item_size, uint32_t start, uint32_t len) {
 	struct __gr_vec_hdr *hdr = __gr_vec_hdr(vec);
-	size_t end = start + len;
+	uint32_t end = start + len;
 
 	if (hdr == NULL || start >= hdr->len || len == 0 || item_size == 0)
 		abort();
@@ -82,8 +82,9 @@ static inline void __gr_vec_del_range(void *vec, size_t item_size, size_t start,
 }
 
 // (internal) delete multiple items stating at a given index
-static inline void *__gr_vec_shift_range(void *vec, size_t item_size, size_t start, size_t len) {
-	size_t end = start + len;
+static inline void *
+__gr_vec_shift_range(void *vec, size_t item_size, uint32_t start, uint32_t len) {
+	uint32_t end = start + len;
 
 	if (len == 0 || item_size == 0 || start > gr_vec_len(vec))
 		abort();
@@ -151,10 +152,10 @@ static inline void *__gr_vec_clone(const void *vec, size_t item_size) {
 
 // iterate over a vector, dereferencing each item into a local variable
 #define gr_vec_foreach(x, v)                                                                       \
-	for (size_t __i = 0, __next = 1; __next && __i < gr_vec_len(v); __next = !__next, __i++)   \
+	for (uint32_t __i = 0, __next = 1; __next && __i < gr_vec_len(v); __next = !__next, __i++) \
 		for (x = v[__i]; __next; __next = !__next)
 
 // iterate over a vector, referencing each item into a local pointer
 #define gr_vec_foreach_ref(p, v)                                                                   \
-	for (size_t __i = 0, __next = 1; __next && __i < gr_vec_len(v); __next = !__next, __i++)   \
+	for (uint32_t __i = 0, __next = 1; __next && __i < gr_vec_len(v); __next = !__next, __i++) \
 		for (p = &v[__i]; __next; __next = !__next)
